fix thread pool instantiation
